Update 3
Added damage and a basic Dummy Enemy Script.

Enemy acts as a Abstract class we can inherit for all enemies. It has all the core stats we need to make an enemy.

All we have to do now to make an enemy is just inherit from Enemy and then in the Child Script ad any nessacry behaviors like movement and so on.