EntryStage
I just designed the possible entry stage for the game. This is mainly a proof of concept rather than a finished product and should be treated as such. I also hid the player, crosshair, and charge indicator within the hierarchy for ease of use when making the level, so unhide it whenever playtesting.